wrote tests for the useForm hook
had to rewrite the custom useForm hook for the following reason: say 'error' was not part of the initial form that is passed into useForm.  but at some point we add 'error' to the form object, when we reset the form that new 'error' key will still be there.  on reset we want the form to be exactly the same as the form that was passed in.  in order to do this the reset has to use/go through the dispatch function